tink-server: add an envar to control running migrations
Two steps for migrations is a pain when we just want to get up and get
running. Two steps is a nice option in some deployment scenarios, but
will likely end up going away all together in the future.

This is setup to be fully backwards compatible by preferring
ONLY_MIGRATION=true over MIGRATE=false.